+++ packages/orq-rc/RELEASES.md
@@ -4799,3 +4799,13 @@ Based on:
 - [typescript v4.0.0-rc.9] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.9] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.9 - packages/orq-rc
+
+## 2025-11-05 15:08:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.648.1 (2.739.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.10] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.10] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.10 - packages/orq-rc